requirements.txt
faiss-cpu
langchain
langchain-community
langchain-google-genai
google-generativeai
